refactor(popup-menu): split deep-search utils by section; rename dir to data-first
