+++ .changeset/data-first-module-move.md
@@ -1,0 +1,5 @@
+---
+'@bazza-ui/react': patch
+---
+
+Internal: the data-first popup-menu pipeline moved from `internal/popup-menu/deep-search/` to `internal/popup-menu/data-first/` and its utilities were split into per-section modules. No public API changes.
