+++ packages/react/src/internal/popup-menu/components/empty/empty.tsx
@@ -8,7 +8,7 @@ import {
   getSlotAttribute,
   useMaybeComponentName,
 } from '../../contexts/component-name-context.js'
-import { useMaybeAsyncMenuCoordinator } from '../../deep-search/async-coordinator.js'
+import { useMaybeAsyncMenuCoordinator } from '../../data-first/async-coordinator.js'
 import { PopupMenuEmptyDataAttributes } from './empty.data-attrs.js'
 
 // Empty doesn't have any state - using an empty object type
