+++ packages/react/src/internal/popup-menu/components/list/list.tsx
@@ -18,8 +18,8 @@ import { useFocusOwner } from '../../contexts/focus-owner-context.js'
 import { usePopupMenuContext } from '../../contexts/popup-menu-context.js'
 import { useMaybeSubmenuContext } from '../../contexts/submenu-context.js'
 import { useMaybeSubpageContext } from '../../contexts/subpage-context.js'
-import { useMaybeDataSurfaceContext } from '../../deep-search/context.js'
-import { DataListInner } from '../../deep-search/data-list.js'
+import { useMaybeDataSurfaceContext } from '../../data-first/context.js'
+import { DataListInner } from '../../data-first/data-list.js'
 import { usePopupMenuKeyboard } from '../../hooks/use-popup-menu-keyboard.js'
 import { PopupMenuListCssVars } from './list.css-vars.js'
 import { PopupMenuListDataAttributes } from './list.data-attrs.js'
