+++ packages/react/src/internal/popup-menu/data-first/index.ts
@@ -1,5 +1,5 @@
 // ============================================================================
-// Deep Search Module
+// Data-First Module
 // ============================================================================
 // Data-first API for popup menus with deep search support.
