+++ packages/react/src/internal/popup-menu/menu-tree/types.ts
@@ -1,4 +1,4 @@
-import type { NodeDef } from '../deep-search/types.js'
+import type { NodeDef } from '../data-first/types.js'
 
 export type PopupMenuIdScope = 'menu' | 'surface'
