+++ packages/react/src/loaders/vanilla.tsx
@@ -11,7 +11,7 @@ import type {
   NodeDef,
   QueryLoaderConfig,
   StaticLoaderConfig,
-} from '../internal/popup-menu/deep-search/types.js'
+} from '../internal/popup-menu/data-first/types.js'
 
 // ============================================================================
 // Vanilla Async Adapter (no external dependencies)
